+++ src/Source.h
@@ -21,119 +21,15 @@ public:
       const char* end,
       int n,
       bool skipEmptyRows = true,
-      const std::string& comment = "") {
-    bool hasComment = comment != "";
-    bool isComment = false, lineStart = true;
-    bool isQuote = false;
-    char openQuote = '\0';
+      const std::string& comment = "");
 
-    const char* cur = begin;
+  const char* skipLine(const char* begin, const char* end, bool isComment);
 
-    for (; n > 0 && cur != end; ++cur) {
-      if (lineStart) {
-        isComment = hasComment && inComment(cur, end, comment);
-      }
-
-      // This doesn't handle escaped quotes or more sophisticated things, but
-      // will work for simple cases.
-      if (!isComment &&
-          (*cur == openQuote ||
-           (openQuote == '\0' && (*cur == '"' || *cur == '\'')))) {
-        isQuote = !isQuote;
-        if (openQuote == '\0') {
-          openQuote = *cur;
-        } else {
-          openQuote = '\0';
-        }
-
-        lineStart = false;
-        continue;
-      }
-
-      if (isComment || isQuote) {
-        continue;
-      }
-
-      if (*cur == '\n' || *cur == '\r') {
-        --n;
-        advanceForLF(&cur, end);
-        ++skippedRows_;
-        lineStart = true;
-      } else if (lineStart) {
-        lineStart = false;
-      }
-    }
-
-    // Skip any more trailing empty rows or comments
-    while (cur < end &&
-           ((skipEmptyRows && (*cur == '\n' || *cur == '\r')) ||
-            (isComment = hasComment && inComment(cur, end, comment)))) {
-      if (isComment) {
-        // skip the rest of the line until the newline
-        while (cur < end && !(*cur == '\n' || *cur == '\r')) {
-          ++cur;
-        }
-      }
-      advanceForLF(&cur, end);
-      ++cur;
-      ++skippedRows_;
-    }
-
-    return cur;
-  }
+  const char* skipDoubleQuoted(const char* begin, const char* end);
 
   size_t skippedRows() { return skippedRows_; }
 
-  static const char* skipBom(const char* begin, const char* end) {
-
-    /* Unicode Byte Order Marks
-       https://en.wikipedia.org/wiki/Byte_order_mark#Representations_of_byte_order_marks_by_encoding
-
-       00 00 FE FF: UTF-32BE
-       FF FE 00 00: UTF-32LE
-       FE FF:       UTF-16BE
-       FF FE:       UTF-16LE
-       EF BB BF:    UTF-8
-   */
-
-    switch (begin[0]) {
-    // UTF-32BE
-    case '\x00':
-      if (end - begin >= 4 && begin[1] == '\x00' && begin[2] == '\xFE' &&
-          begin[3] == '\xFF') {
-        return begin + 4;
-      }
-      break;
-
-    // UTF-8
-    case '\xEF':
-      if (end - begin >= 3 && begin[1] == '\xBB' && begin[2] == '\xBF') {
-        return begin + 3;
-      }
-      break;
-
-    // UTF-16BE
-    case '\xfe':
-      if (end - begin >= 2 && begin[1] == '\xff') {
-        return begin + 2;
-      }
-      break;
-
-    case '\xff':
-      if (end - begin >= 2 && begin[1] == '\xfe') {
-
-        // UTF-32 LE
-        if (end - begin >= 4 && begin[2] == '\x00' && begin[3] == '\x00') {
-          return begin + 4;
-        }
-
-        // UTF-16 LE
-        return begin + 2;
-      }
-      break;
-    }
-    return begin;
-  }
+  static const char* skipBom(const char* begin, const char* end);
 
   static SourcePtr create(Rcpp::List spec);